backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pymongo>=3.12.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3
python-dotenv>=0.19.0
pytz>=2021.1